cargo test runs all_tests
