+++ Cargo.toml
@@ -15,7 +15,3 @@ clap = { version = "4.4.6", features = ["cargo", "derive", "env"] }
 log = "0.4"
 simplelog = { version = "0.12", features = ["paris"] }
 tokio = "1.33.0"
-
-[[bin]]
-name = "all_tests"
-path = "src/scripts/all_tests.rs"
